requirements.txt
--index-url http://dist.plone.org/thirdparty/
Django==1.3.1
PIL==1.1.7
PyX==0.11.1
argparse==1.2.1
coverage==3.5.1
django-haystack==1.2.6
psycopg2==2.4.4
wsgiref==0.1.2
